refactor-dice-string-parser-create - Moved from test_dice_string_parser.py to test_default_dice_string_parser.py
